Add documentation links to migration-related banners
