+++ packages/app/src/cli/services/context/breakdown-extensions.ts
@@ -20,10 +20,10 @@ import {ExtensionSpecification} from '../../models/extensions/specification.js'
 import {rewriteConfiguration} from '../app/write-app-configuration-file.js'
 import {AppConfigurationUsedByCli} from '../../models/extensions/specifications/types/app_config.js'
 import {removeTrailingSlash} from '../../models/extensions/specifications/validation/common.js'
+import {throwUidMappingError} from '../../prompts/uid-mapping-error.js'
 import {deepCompare, deepDifference} from '@shopify/cli-kit/common/object'
 import {encodeToml} from '@shopify/cli-kit/node/toml'
 import {zod} from '@shopify/cli-kit/node/schema'
-import {AbortError} from '@shopify/cli-kit/node/error'
 
 export interface ConfigExtensionIdentifiersBreakdown {
   existingFieldNames: string[]
@@ -68,15 +68,7 @@ export async function extensionsIdentifiersDeployBreakdown(options: EnsureDeploy
     const unMigratedextensions = remoteExtensionsRegistrations.app.extensionRegistrations.filter((ext) => !ext.id)
 
     if (unMigratedextensions.length > 0) {
-      const message = ['Your app has extensions which need to be assigned', {command: 'uid'}, 'identifiers.']
-      const nextSteps = [
-        'You must first map IDs to your existing extensions by running',
-        {command: 'shopify app deploy'},
-        'interactively, without',
-        {command: '--force'},
-        'to finish the migration.',
-      ]
-      throw new AbortError(message, nextSteps)
+      throwUidMappingError()
     }
   }
